Refactor current tests
